refactor: Store 파일명 재정의
