+++ src/components/common/CommentBottomSheet/GlobalCommentBottomSheet.tsx
@@ -5,8 +5,8 @@ import ReplyList from '@/components/common/Post/ReplyList';
 import { getComments, type CommentData } from '@/api/comments/getComments';
 import { postReply } from '@/api/comments/postReply';
 import { useReplyActions } from '@/hooks/useReplyActions';
-import { useReplyStore } from '@/stores/useReplyStore';
-import { useCommentBottomSheetStore } from '@/stores/useCommentBottomSheetStore';
+import { useReplyStore } from '@/stores/replyStore';
+import { useCommentBottomSheetStore } from '@/stores/commentBottomSheetStore';
 import { usePopupActions } from '@/hooks/usePopupActions';
 import { getRoomPlaying } from '@/api/rooms/getRoomPlaying';
 import { isRoomCompleted } from '@/utils/roomStatus';
