+++ src/pages/feed/FeedDetailPage.tsx
@@ -12,7 +12,7 @@ import { useReplyActions } from '@/hooks/useReplyActions';
 import { getFeedDetail, type FeedDetailData } from '@/api/feeds/getFeedDetail';
 import { getComments, type CommentData } from '@/api/comments/getComments';
 import { deleteFeedPost } from '@/api/feeds/deleteFeedPost';
-import { useReplyStore } from '@/stores/useReplyStore';
+import { useReplyStore } from '@/stores/replyStore';
 
 const FeedDetailPage = () => {
   const navigate = useNavigate();
